Add files via upload
The gnomad files that are used as a base for the comparisons. They provide the population statistics useful for analysis of structural variants. the barebones file has just the information Truvari needs to compare the variants, the truv file has all of the population statistic information, and the AF file has the AF, pop_max_frew, and num_homalts.